Revert "feat: Update middleware for authentication and error handling."
This reverts commit 00beec8f257312a065e5cc5d027e0314ae4e29f2.
